fix(lint): 禁用 revive 吵闹规则 + 修复 user G115 溢出
- .golangci.yaml: disable exported + package-comments revive rules
  (stutters/comment 噪声 50+ 项 → 0)
- user_model.go: ID uint64 → int64 (与领域层一致)
- user_repository.go: 移除 uint64() 强制转换

+++ internal/infrastructure/persistence/gorm/user_model.go
@@ -4,7 +4,7 @@ import "time"
 
 // UserModel is the GORM database model for users.
 type UserModel struct {
-	ID           uint64    `gorm:"primaryKey;column:id"`
+	ID           int64    `gorm:"primaryKey;column:id"`
 	Username     string    `gorm:"size:50;not null;uniqueIndex;column:username"`
 	PasswordHash string    `gorm:"size:255;not null;column:password_hash"`
 	CreatedAt    time.Time `gorm:"column:created_at"`
@@ -16,7 +16,7 @@ func (UserModel) TableName() string { return "users" }
 
 func (m *UserModel) toEntity() *userEntity {
 	return &userEntity{
-		ID: int64(m.ID), Username: m.Username, PasswordHash: m.PasswordHash,
+		ID: m.ID, Username: m.Username, PasswordHash: m.PasswordHash,
 		CreatedAt: m.CreatedAt, UpdatedAt: m.UpdatedAt,
 	}
 }
